--- a/polars/polars-core/src/fmt.rs
+++ b/polars/polars-core/src/fmt.rs
@@ -616,7 +616,7 @@ impl<T> FmtList for ObjectChunked<T> {
     }
 }
 
-#[cfg(all(test, feature = "temporal"))]
+#[cfg(all(test, feature = "temporal", feature="dtype-date32", feature = "dtype-date64"))]
 mod test {
     use crate::prelude::*;
     use polars_arrow::prelude::PrimitiveArrayBuilder;