--- a/polars/polars-core/src/vector_hasher.rs
+++ b/polars/polars-core/src/vector_hasher.rs
@@ -38,6 +38,7 @@ impl Default for IdHasher {
 
 pub(crate) type IdBuildHasher = BuildHasherDefault<IdHasher>;
 
+#[derive(Debug)]
 pub(crate) struct IdxHash {
     // idx in row of Series, DataFrame
     pub(crate) idx: u32,