@ritchie46 2021-04-03
Some benchmarks:

Improvement of comparison functions

```
before               time:   [1.1846 us 1.1847 us 1.1848 us]      

after                 time:   [911.62 ns 911.84 ns 912.09 ns]                      
               change: [-21.535% -21.457% -21.350%] (p = 0.00 < 0.05)
```

Effect of this on groupby `q6` and `q9`

```
before
groupby q6              time:   [41.514 ms 41.850 ms 42.207 ms]                       
groupby q9              time:   [102.27 ms 102.75 ms 103.26 ms]                       


after
groupby q6              time:   [39.128 ms 39.505 ms 39.891 ms]                       
                        change: [-6.7812% -5.6037% -4.4217%] (p = 0.00 < 0.05)
                        Performance has improved.
groupby q9              time:   [98.680 ms 99.229 ms 99.805 ms]                       
                        change: [-4.1114% -3.4288% -2.6237%] (p = 0.00 < 0.05)
                        Performance has improved.
```